app_users/views.py
from rest_framework import viewsets

from app_users.models import User
from app_users.serializers import UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    serializer_class = UserSerializer
    queryset = User.objects.all()
